rename detectors
