+++ mmpose/models/detectors/__init__.py
@@ -1,4 +1,4 @@
-from .bottom_up import BottomUp
+from .associative_embedding import AssociativeEmbedding
 from .interhand_3d import Interhand3D
 from .mesh import ParametricMesh
 from .multi_task import MultiTask
@@ -6,6 +6,6 @@ from .pose_lifter import PoseLifter
 from .top_down import TopDown
 
 __all__ = [
-    'TopDown', 'BottomUp', 'ParametricMesh', 'MultiTask', 'PoseLifter',
-    'Interhand3D'
+    'TopDown', 'AssociativeEmbedding', 'ParametricMesh', 'MultiTask',
+    'PoseLifter', 'Interhand3D'
 ]
